Fix extension type

+++ Sources/CorrelationVector/AtomicHelper.h
@@ -3,9 +3,10 @@
 
 #import <stdatomic.h>
 #import <stdbool.h>
+#import <stdint.h>
 
 /**
  Compare and exchange contained value.
  Automatic binding to swift doesn't work due to the bug - https://openradar.appspot.com/27161329
  */
-bool compareAndSwap(_Atomic(long) *value, long * expected, long desired);
+bool compareAndSwap(_Atomic(uint32_t) *value, uint32_t *expected, uint32_t desired);
