+++ Sources/CorrelationVector/AtomicHelper.m
@@ -3,6 +3,6 @@
 
 #import "AtomicHelper.h"
 
-bool compareAndSwap(_Atomic(long) *value, long * expected, long desired) {
+bool compareAndSwap(_Atomic(uint32_t) *value, uint32_t *expected, uint32_t desired) {
   return atomic_compare_exchange_strong(value, expected, desired);
 }
